Commit 100%

+++ src/main/java/lang/print/gaps/finalModuleTask/DivisionByZero.java
@@ -2,7 +2,7 @@ package lang.print.gaps.finalModuleTask;
 
 public class DivisionByZero {
     public static void main(String[] args) {
-        int number = 10;
+        float number = 10;
         System.out.println((float)number/0);
     }
 }
